DeleteMapping

+++ src/main/java/boot/crud/controller/UserController.java
@@ -5,10 +5,9 @@ import boot.crud.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Controller
@@ -39,9 +38,13 @@ public class UserController {
         return "redirect:/users";
     }
 
-    @GetMapping("/user-delete/{id}")
+    @DeleteMapping("/user-delete/{id}")
     public String deleteUser(@PathVariable("id") Long id) {
-        userService.deleteById(id);
+        if (userService.findById(id) != null) {
+            userService.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("no such user with id " + id);
+        }
         return "redirect:/users";
     }
